requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.30
aiomysql>=0.2.0
httpx>=0.27.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
pydantic-settings>=2.2.1
